ADD : postdetail toggle

+++ src/pages/PostDetail.jsx
@@ -1,6 +1,5 @@
-import React from "react";
-import Header from "../components/Header";
-
+import React , {useState}from "react";
+import styled from "styled-components";
 import { useNavigate, useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import CommentCreate from "../components/CommentCreate";
@@ -12,6 +11,7 @@ import {
 } from "../redux/modules/PostDetailsSlice";
 import Layout from "../components/Layout"
 import Footer from "../components/Footer"
+import back from "../assets/back.png" 
 
 const PostDetail = () => {
   const navigate = useNavigate();
@@ -28,20 +28,44 @@ const PostDetail = () => {
 
   //게시글 삭제
   const onDeleteHandler = (payload) => {
-    dispatch(__deletePost(payload));
+    if(window.confirm("정말 삭제하시겠습니까?")){
+      dispatch(__deletePost(payload));
     window.location.replace("/postread/all");
+    }else{
+      window.location.reload()
+    }
+    
     //   navigate("/postread/all");
   };
+
+  const [editTg, setEidtTg] = useState({
+    isEdit:false,
+  });
+  
+  const editToggleHandler = (postId) => {
+    const newEdit = {
+      isEdit:!editTg.isEdit,
+    }
+    setEidtTg(newEdit)
+  }
+
 
   return (
     <>
     <Layout>
-      <Header />
-      <div>PostDetail</div>
+      <EditHead>
+      <div><img onClick={()=> {navigate(-1)}} style={{width:25, height : 25}} src={back}/></div>
+      <Tgbutton onClick={editToggleHandler}>···</Tgbutton>
+                {editTg.isEdit === true ? (
+                <ToggleNav>
+                  <Button onClick={() => navigate(`/postupdate/${params.id}`)}>수정</Button>
+                  <Button onClick={() => {onDeleteHandler(params.id);}}>글삭제</Button>
+                </ToggleNav>
+                ):null 
+                } 
+    </EditHead>
       <button onClick={() => onCartButton(posts.postId)}>찜</button>
-      <button onClick={() => navigate(`/postupdate/${params.id}`)}>
-        수정하기
-      </button>
+      
       <div>
         <div>{posts.title}</div>
         {posts.images !== undefined &&
@@ -61,13 +85,7 @@ const PostDetail = () => {
 
       <div>
         <button onClick={() => navigate(-1)}>이전으로</button>
-        <button
-          onClick={() => {
-            onDeleteHandler(params.id);
-          }}
-        >
-          글삭제
-        </button>
+        
         <Footer/>
       </div>
       </Layout>
@@ -76,3 +94,36 @@ const PostDetail = () => {
 };
 
 export default PostDetail;
+
+// 수정 삭제 토글 및 뒤로가기
+const EditHead = styled.div`
+position : relative;
+display: flex;
+justify-content: space-between;
+padding: 10px;
+`
+
+const Tgbutton = styled.button`
+  border:none;
+  font-weight:600;
+  width:50px;
+  background-color:white;
+`
+const ToggleNav = styled.div`
+  width : 50px;
+  height: 80px;
+  position: absolute;
+  right : 10px;
+  top : 50px;
+  `
+const Button = styled.button`
+  width:50px;
+  height:40px;
+  margin-bottom:3px;
+  border:1px solid #ddd;
+  border-radius:5px;
+  background-color:#fff;
+  &:hover {
+    background-color:red;
+  }
+`
